Update NFTProperties.tsx
On behalf of @dkackman:
For an audio NFT (or at least the one indicated in the #1786)

rawValue on line 44 of \chia-blockchain-gui\packages\gui\src\components\nfts\NFTProperties.tsx can be undefined. Added a simple check alongside where typeof object is checked and return null. Haven't seen downstream consequences and the nft detail page now opens without error.

value on that line could also be set to an empty string or other default but I don't know enough about the code to know what is the best approach so deiced an early bail out, since it already is in there, was safest.

Co-Authored-By: Don Kackman <dkackman@gmail.com>