Remove default empty fee and change condition for adding a fee.
The `canAdd` condition for `OfferBuildFeeSection` didn't allow to add a fee when viewing an offer that is not yet accepted. Also, the `!fields.length` (which means that there should be no existing fee in order to add a fe) condition should always be `true` in order to be able to add a fee. This is because fee is an array but we should only add 1 fee.

If a fee was given, it would show it, but if we remove the fee it would not be allowed to add a fee. This was the case for `TakeOffer`. With the new code that allow to add a fee for `CreateOffer`, we now have the same behaviour :

- Fee is optional.
- If fee is not given, the user can add a fee manually
- If fee is given the user can edit it.
- If fee is given and the user remove the fee by accident, he can still add a fee again.